+++ server/modules/monitor/mariadbmon/server_utils.hh
@@ -255,7 +255,7 @@ public:
     int64_t     received_heartbeats = 0;                /* How many heartbeats the connection has
                                                          * received */
 
-    int seconds_behind_master = mxs::Target::RLAG_UNDEFINED;    /* How much behind the slave is. */
+    int64_t seconds_behind_master = mxs::Target::RLAG_UNDEFINED;    /* How much behind the slave is. */
 
     /* Time of the latest gtid event or heartbeat the slave connection has received, timed by the monitor. */
     maxbase::TimePoint last_data_time = maxbase::Clock::now();
